+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.8 (TBA)
 
 * Removed HTTPoison and use `:httpc` instead
+* Fixed bug where HTML with no style tags resulted in all existing inline styles being removed
 
 ## v0.3.7 (2019-07-05)